Testing of Functions, Q28_4 and PointWebMercator classes is done. Documentation has been completed.

+++ src/ch/epfl/javelo/Q28_4.java
@@ -12,17 +12,12 @@ public final class Q28_4 {
     // Non-instantiable class
     private Q28_4() {}
 
-    ;
 
     /**
-     * TODO:
-     * Naive implementation of ofInt, I need to handle those two cases:
-     *  - If negative value, avoid change on the sign. (can we realistically have a negative value as an input ?)
-     *  - If i << 4 is bigger than Integer.MAX_VALUE // Is it a case that we need to worry about ? I don't think so.
-     *
      * @param i the int to format into Q28_4
      * @return The Q28_4 byte string that has the same value as i.
      */
+
     public static int ofInt(int i) {
         return i << 4;
     }
@@ -35,7 +30,7 @@ public final class Q28_4 {
      */
 
     public static double asDouble(int q28_4) {
-        return Math.scalb((double) q28_4, -4);
+        return Math.scalb(q28_4, -4);
     }
 
     /**
